MarionL -  creating projects needs user
Adding an initial project_leader when a project is created. This means
that a user is now required in order to create projects from the DAO
(and later the REST). Modified the unit tests accordingly.

Similarly in order to deleteProject you need to delete any members
associated to the project first.