MarionL - changed db url to use remote url.
Changed db url to use remote url (Amazon AWS) - in the mysql
configuration.
Made the same remote url change to the example server.xml file for
tomcat.
Modified gitignore to allow war file to be checked in - this is the
reason it didn't show up during final master check-in.
Added the war file.